+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -14,11 +14,9 @@ import static org.mockito.Mockito.doThrow;
 import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.List;
-import kitchenpos.menu.application.MenuGroupService;
-import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.validator.MenuValidator;
 import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 import kitchenpos.menu.dto.request.MenuProductRequest;
 import kitchenpos.menu.dto.request.MenuRequest;
 import kitchenpos.menu.dto.response.MenuResponse;
