+++ src/test/java/kitchenpos/order/ui/OrderRestControllerTest.java
@@ -3,7 +3,7 @@ package kitchenpos.order.ui;
 import static kitchenpos.table.ui.TableRestControllerTest.좌석_생성_요청;
 import static kitchenpos.menu.ui.MenuGroupRestControllerTest.메뉴_그룹_생성_요청;
 import static kitchenpos.menu.ui.MenuRestControllerTest.메뉴_생성_요청;
-import static kitchenpos.ui.ProductRestControllerTest.상품_생성_요청;
+import static kitchenpos.product.ui.ProductRestControllerTest.상품_생성_요청;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
@@ -20,10 +20,10 @@ import kitchenpos.menu.dto.request.MenuProductRequest;
 import kitchenpos.menu.dto.request.MenuRequest;
 import kitchenpos.order.dto.request.OrderRequest;
 import kitchenpos.order.dto.request.OrderStatusRequest;
-import kitchenpos.dto.request.ProductRequest;
+import kitchenpos.product.dto.request.ProductRequest;
 import kitchenpos.menu.dto.response.MenuResponse;
 import kitchenpos.order.dto.response.OrderResponse;
-import kitchenpos.dto.response.ProductResponse;
+import kitchenpos.product.dto.response.ProductResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.core.ParameterizedTypeReference;
