+++ src/test/java/kitchenpos/product/domain/ProductTest.java
@@ -1,13 +1,11 @@
-package kitchenpos.domain;
+package kitchenpos.product.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.math.BigDecimal;
-import kitchenpos.exception.KitchenposException;
+import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.NullSource;
 
 class ProductTest {
     @Test
